+++ R/search.R
@@ -3,18 +3,17 @@
 #' @param x A character string to find in GHO labels.
 #' @param dimension A GHO dimension where codes will be searched.
 #' @param gho An object from \code{\link{get_gho_dimensions}} or
-#'  \code{\link{get_gho_codes}}.
+#'  \code{\link{get_gho_values}}.
 #'
-#' @return A vector of dimensions or codes whose labels mach the search string.
-#'  Labels are given as as a \code{label} attribute.
+#' @return A `GHO` object
 #'
 #' @export
 #'
 #' @examples
 #'
-#' search_codes("neonatal", dimension = "GHO")
+#' search_values("neonatal", dimension = "GHO")
 #'
-#' result <- get_gho_codes(dimension = "REGION")
+#' result <- get_gho_values(dimension = "REGION")
 #' search_gho(result, "asia")
 #'
 search_gho <- function(gho, x) {
@@ -30,6 +29,6 @@ search_dimensions <- function(x) {
 
 #' @rdname search_gho
 #' @export
-search_codes <- function(x, dimension = "GHO") {
-  search_gho(get_gho_codes(dimension), x)
+search_values <- function(x, dimension = "GHO") {
+  search_gho(get_gho_values(dimension), x)
 }